script: Move finalizer implementations to script_bindings.
